refactor(tests): refactorizar tests de módulo Queues - dividir queue.service.spec.ts
- Dividir queue.service.spec.ts (839 líneas) en 5 archivos especializados
- Crear queue.service.core.spec.ts (271 líneas): tests de adición de jobs
- Crear queue.service.metrics.spec.ts (264 líneas): tests de métricas
- Crear queue.service.management.spec.ts (265 líneas): tests de gestión
- Crear queue.service.lifecycle.spec.ts (320 líneas): tests de ciclo de vida
- Crear queue.service.edge-cases.spec.ts (354 líneas): tests edge cases
- Crear helpers/queue-service.test-helpers.ts: factories y assertions
- Eliminar archivo original queue.service.spec.ts

Helpers creados:
- createMockQueue(): factory para mocks de Bull Queue
- createMockJobCounts(): factory para job counts
- expectValidQueueMetrics(): assertion de estructura de métricas
- expectEventListeners(): assertion de event listeners
- getEventListener(): helper para obtener listeners

Tests: 91 passing (100%), Promedio: 295 líneas/archivo (antes: 839)
